Fix air time display limit

+++ src/openrct2/core/UnitConversion.h
@@ -22,6 +22,6 @@ namespace OpenRCT2
     uint8_t MetresToBaseZ(int16_t metres);
     int32_t HeightUnitsToMetres(int32_t heightUnit);
     int32_t ToHumanReadableSpeed(int32_t baseSpeed);
-    uint16_t ToHumanReadableAirTime(uint16_t airTime);
+    int32_t ToHumanReadableAirTime(uint16_t airTime);
     int32_t ToHumanReadableRideLength(int32_t rideLength);
 } // namespace OpenRCT2
